Create github workflow for dev commits
Dev should still only be used occasionally, not for working on features. It's more of a staging ground to put newly developed features or improvements from other branches on, to make sure that it doesn't break the project. Once the dev branch is verified to work, it can be attempted to merge with main.

Signed-off-by: Alpha-Craft <77483478+AlphaCraft9658@users.noreply.github.com>